### 2021-11-02 #3879 @whedon
[PRE REVIEW]: mmrefpoints: Projecting long-term marine mammal abundance with bycatch
**Submitting author:** @mcsiple (<a href="http://orcid.org/0000-0002-4260-9856">Margaret Siple</a>)
**Repository:** <a href="https://github.com/mcsiple/mmrefpoints" target ="_blank">https://github.com/mcsiple/mmrefpoints</a>
**Version:** v0.1.1
**Reviewers:** @ha0ye and @eirenjacobson
