### 2021-11-04 #3879 @fboehm
@crvernon, @eirenjacobson, @neerajdhanraj, @tbrown122387, @ha0ye   - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
