### 2021-11-04 #3879 @crvernon
@fboehm Apologies, I am currently the editor for several other JOSS articles in-review and will not have time to be a reviewer on this one.  Looks like a good one though!   Kind regards!
