### 2021-11-04 #3888 @whedon
[REVIEW]: mmrefpoints: Projecting long-term marine mammal abundance with bycatch
**Submitting author:** @mcsiple (<a href="http://orcid.org/0000-0002-4260-9856">Margaret Siple</a>)
**Repository:** https://github.com/mcsiple/mmrefpoints
**Branch with paper.md** (empty if default branch):
**Version:** v 1.0.1
