### 2021-11-17 #3888 @fboehm
Hi, @eirenjacobson - sorry for my error. Can you check to see if you can check the boxes in the checklist? Just try to check one, to see if the checkmark appears, then uncheck it, assuming that you haven't actually verified that piece of information for the review. I think that there was an issue with your invitation, but the bot whedon may have resolved it. Thanks again!
