### 2021-12-02 #3888 @eirenjacobson
Hi all, I had planned to complete this review this week but I am currently participating in strike action to improve UK higher education staff pension, pay, equality and working conditions (for more information, see https://www.ucu.org.uk/article/11896/Why-were-taking-action). I will try to complete my review when I return from strike next week.
