### 2021-12-02 mcsiple/mmrefpoints#54 @ha0ye
JOSS Review Notes (Hao Ye, @ha0ye)
I'm collecting all the details for my review in this one issue. @mcsiple or another maintainer may consider splitting specific action items into their own separate issues for better task management.

Note: many of these are suggestions for improvements, and in some cases are not actual obstacles for passing the review process at JOSS. These items are indicated by `[~]` rather than an empty box `[ ]`. I've also excised all the checked boxes from my review over in https://github.com/openjournals/joss-reviews/issues/3888

  * Yes, but `DESCRIPTION` should be updated to be in line with the author list and version number in CITATION.CFF.
  * .zenodo.json has a different author list than the entry at Zenodo: https://zenodo.org/record/4758402 - The repository owner might want to enable auto-archiving in Zenodo for GitHub releases - https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content

  * warning on install:
  > ─  excluding invalid files
  >    Subdirectory 'R' contains invalid file names:
  >      ‘_disable_autoload.R’

  Could this be added to `.Rbuildignore`? I'm not sure if it is present to help with running the shiny app as a developer, and would be safe to exclude from the built package for end-users.

  * I think `arrange` needs to be added to the functions to importFrom dplyr on line 15 of `17_plot_proj.R` and line 20 of `18_multiplot_proj.R`
  * For the model projection plot, it might be useful to use one of the interactive plotting packages like plotly, which would let the user zoom in on certain portions (to better see distinction between trajectories or better identify when extirpation is predicted). In addition, the legend for Bycatch level only has the labels for "{High end / Midpoint / Low end} of bycatch range", and it may be useful to calculate the numbers from the specified bycatch mortality range to make the legend more informative by itself.
  * When toggling between bycatch mortality by "Numbers per year" and "Mortality rate", the interface for selecting Bycatch CV changes between a text box and a slider. Is there a reason for this?
  * The kite plot in the "Advanced Projections" seems to occasionally error out with a "need finite 'ylim' values" message. (e.g. selecting "Sea lion", leaving all parameters at default values, and clicking on "Run projections")
  * The population time series plot seems to fail when both "Show individual projections" and "Show multiple depletion levels" is checked.
  * It'd be ideal to reorder the bycatch level factor so that Low, Midpoint and High are in either increasing or decreasing order.
  * The violin plots are flat for some bycatch levels, resulting in horizontal lines instead of a distribution. There might be a way to indicate such, and perhaps to better show grouping for "bycatch level" within "years after projections start"
  * The design is a bit awkward with the side-by-side of parameters and projection plots. I feel as though I would prefer to have the usage steps proceed from top to bottom, so that I set parameters in a section, then scroll down and click to run projections, then scroll down and see model outputs.
  * You have different kinds of summary outputs for the model projections. It might be nice to add an option to download the raw projection data for someone to do their own calculations.
  * The GAMMS link in the "PBR & PBR calculator" tab doesn't seem to work.
  * It does seem a bit confusing that the input arguments to `dynamics()` have separate arguments for the life history parameters, whereas in `projections()`, they're given as a named list. I think this could be made more consistent, and that the list structure is probably better as it is easier to refer to as a single object to pass to multiple functions.

  * `DESCRIPTION` has a duplicated `LazyData` line that is preventing building the package.
  * There are a few points where you have `dplyr::arrange(dplyr::desc(name, sim))` which produces an error. I think it should be something like `dplyr::arrange(dplyr::desc(name), dplyr::desc(sim))`
  * Is this possible to do? I'm not sure about Shiny limitations.
    > For the "Solve for Bycatch" tab, it would be ideal to show all the life history and other model parameters from the Advanced tab here, just so someone using it can confirm that the numbers are correct without having to switch tabs. I also think showing the table of tested bycatch rate values would be good, as it would allow people to then go back to the Advanced tab and test/confirm the results.

  * Yes, in the readme. The application does contain some functionality to output a report from the "PBR & PBR calculator" tab. I don't know how useful it would be for an end-user to be able to go through model projections in the Shiny app and then output a reproducible R Markdown report, but that could be a consideration as a feature. (not necessary here, but perhaps if stakeholder feedback shows it as a useful inclusion)

  * There seems to be some duplication in the "Model description" vignette and the "About the project model" tab of the shiny app. There's some concern about keeping the two synchronized if/when updates happen. Is it possible perhaps to embed one of the Rmarkdown files in the other, so that changes only need to happen in one location?
  * I would like to see some additional description in the "Model description" vignette about the function arguments that the variables map to. The names are generally straightforward, but it would help to have that centralized table of variable names in the vignette.

  * The testing infrastructure is present, but is currently a bit bare-bones.
  * It appears that testing of Shiny applications is still a bit in flux. Shiny is now on ver 1.7.0, but doesn't seem to fully have the testing functions that https://engineering-shiny.org/build-yourself-safety-net.html?q=testing#shinytestserver mentions. One option that I've seen in the rOpenSci community is using VCR to test the interactive components - https://github.com/ropensci/vcr/
  * Regardless of the above, tests could be added for the modeling and calculation functions in the package, possibly along with test datasets to confirm correct computation. The R package book contains details about testing here - https://r-pkgs.org/tests.html

  * There is some basic info in the package README:
    - The badge link should point to the issues page for this repo.
    - I don't personally find the linked page in R package book useful for guiding new contributors. (and practiced contributors likely already know of the content or won't need it.)
    - Consider a separate CONTRIBUTING.md file that also links to the Code of Conduct, and provides more specific instructions. There is some general info here - https://mozillascience.github.io/working-open-workshop/contributing/

  * There is some jargon that is specific to those involved in fisheries management, but the target audience is made clear in the Statement of Need, so I do not feel this is a problem in this section of the paper.

  * I found https://github.com/NorskRegnesentral/rSPAMM as another marine mammal assessment package in R. It would be helpful to include a sentence or two on the differences (perhaps the underlying model and intended audience?)

  * Some references that were previously unpublished, are now published and the citation info should be updated, e.g. https://www.frontiersin.org/articles/10.3389/fmars.2021.757330/full
