### 2021-12-02 #3888 @ha0ye
I have completed my review, and my detailed notes are in the issue linked directly above.
