### 2021-12-07 #3879 @fboehm
> @fboehm yeah sure I can help out

Thank you, @tbrown122387 ! I've now added you as a reviewer in the "review" issue, https://github.com/openjournals/joss-reviews/issues/3888. You should find a checklist to guide your review there.
