### 2021-12-11 mcsiple/mmrefpoints#55 @tbrown122387
`magick` is not installed with the library
For some reason my `devtools::install_github()` fails. I get this

```
Installing package into ‘/home/taylor/R/x86_64-pc-linux-gnu-library/4.1’
(as ‘lib’ is unspecified)
ERROR: dependency ‘magick’ is not available for package ‘mmrefpoints’
* removing ‘/home/taylor/R/x86_64-pc-linux-gnu-library/4.1/mmrefpoints’
Warning messages:
1: In i.p(...) : installation of package ‘magick’ had non-zero exit status
2: In i.p(...) :
  installation of package ‘/tmp/RtmpYNeccD/file19fef3d8f047b/mmrefpoints_0.0.0.9000.tar.gz’ had non-zero exit status
```

`magick` is in the `DESCRIPTION` file, so that's not the problem. It attempted to install `magick` but I got this:

```
* installing *source* package ‘magick’ ...
** package ‘magick’ successfully unpacked and MD5 sums checked
** using staged installation
Package Magick++ was not found in the pkg-config search path.
Perhaps you should add the directory containing `Magick++.pc'
to the PKG_CONFIG_PATH environment variable
No package 'Magick++' found
Using PKG_CFLAGS=
Using PKG_LIBS=-lMagick++-6.Q16
--------------------------- [ANTICONF] --------------------------------
Configuration failed to find the Magick++ library. Try installing:
 - deb: libmagick++-dev (Debian, Ubuntu)
 - rpm: ImageMagick-c++-devel (Fedora, CentOS, RHEL)
 - csw: imagemagick_dev (Solaris)
 - brew imagemagick@6 (MacOS)
For Ubuntu versions Trusty (14.04) and Xenial (16.04) use our PPA:
   sudo add-apt-repository -y ppa:cran/imagemagick
   sudo apt-get update
   sudo apt-get install -y libmagick++-dev
If Magick++ is already installed, check that 'pkg-config' is in your
PATH and PKG_CONFIG_PATH contains a Magick++.pc file. If pkg-config
is unavailable you can set INCLUDE_DIR and LIB_DIR manually via:
R CMD INSTALL --configure-vars='INCLUDE_DIR=... LIB_DIR=...'
-------------------------- [ERROR MESSAGE] ---------------------------
<stdin>:1:10: fatal error: Magick++.h: No such file or directory
compilation terminated.
--------------------------------------------------------------------
ERROR: configuration failed for package ‘magick’
* removing ‘/home/taylor/R/x86_64-pc-linux-gnu-library/4.1/magick’
```

`.pc` files are DOS-specific if I recall correctly. I'm running this on

```
Distributor ID:	Ubuntu
Description:	Ubuntu 20.04.3 LTS
Release:	20.04
Codename:	focal
```

openjournals/joss-reviews#3888
