### 2021-12-13 mcsiple/mmrefpoints#55 @mcsiple
Hi @tbrown122387 -- sorry about that.

Have you tried installing the ImageMagick++ library on your machine? For Ubuntu it should be installable using:
`sudo apt-get install libmagick++-dev`
(I got this from [here](https://cran.r-project.org/web/packages/magick/vignettes/intro.html)).

Secondly, does `mmrefpoints` load even though you get this error on install? I only ask because in the package, `magick` is used to run the gif engine for one of the tabs in the app and plays a pretty small role (but one I haven't been able to replace w anything else). If you can still install it, you can review the rest of the package and app's functionality while I figure out the `magick` issue.
