### 2021-12-27 mcsiple/mmrefpoints#55 @tbrown122387
@mcsiple okay great, that works! Maybe a small tidbit should be added to the install section in the README that mentions the added difficulty for Linux folks?

Regarding your second paragraph, I doubt the rest of it would load. I'm looking at the last line of the error message. It just deletes all of the progress if it doesn't get to the end successfully.
