### 2021-12-27 mcsiple/mmrefpoints#56 @tbrown122387
shiny app doesn't run properly
I don't think the Shiny app is working correctly. It opens up something in my web browser on localhost, but it looks like this:

![Screenshot from 2021-12-27 18-19-37](https://user-images.githubusercontent.com/1740324/147513198-e4cae0b1-6db2-4ff5-b7d6-b6c8bef72cae.png)

The console running it in the background throws the following warnings:

```
> mmrefpoints::run_app()
Loading required package: shiny

Listening on http://127.0.0.1:5687
[1] "Language change: en"
Warning: Error in $: $ operator is invalid for atomic vectors
  [No stack trace available]
Warning: Error in $: $ operator is invalid for atomic vectors
  [No stack trace available]
Warning: Error in $: $ operator is invalid for atomic vectors
  [No stack trace available]

```

openjournals/joss-reviews#3888
