### 2021-12-29 88564ae
update readme with install instructions for Linux users
