### 2021-12-29 mcsiple/mmrefpoints#55 @mcsiple
I [added a note](https://github.com/mcsiple/mmrefpoints/commit/88564ae260aa5d03dc0f547ae2d088ee6667925d) to the README for Linux users! Just saw your other message and will answer shortly.
