### 2021-12-29 mcsiple/mmrefpoints#56 @mcsiple
Interesting...I've gotten this error before when I've had issues with the `shiny.i18n` translation package; I'll have to poke around. You're using a Linux machine, right?
