### 2021-12-31 mcsiple/mmrefpoints#56 @mcsiple
I have to download Ubuntu and try to run the app through a Linux VM to test everything out, so it might take a little bit for me to figure out the issue-- if there are other parts of the package you can test out (individual functions, maybe) in the meantime, I think doing those first will be the most efficient way to use our time! Sorry about this; I had not tested it on Linux before.
