### 2022-01-05 mcsiple/mmrefpoints#56 @mcsiple
Sigh... OK this isn't ideal but I get the same error when I install and try to run the app on my Windows machine. For reasons I do not yet understand, it *does not* load when I try:
```
mmrefpoints::run_app()
```
(throws the same error: `$ operator is invalid for atomic vectors`
But it *does* run when I do:

```
library(mmrefpoints)
run_app()
```

Sorry for all the back and forth but I'm just trying to at least get to a spot where you can run the app and comment on it-- and I'll work on this issue in the background.
