### 2022-01-05 ba40b64
add globals.R file
