+++ R/globals.R
@@ -1,0 +1,5 @@
+# Global variables
+# Based on suggestion from https://community.rstudio.com/t/how-to-solve-no-visible-binding-for-global-variable-note/28887
+
+utils::globalVariables(c("lh","Code","S0","S1plus","AgeMat","nages","fmax","lambdaMax","K1plus","Type","dat","BycatchLvl","abund10","abund50","Depl",
+                         "prebuild50","abundrel50","depletion","fishing.rate","probability","S0.w","S1plus.w"))
