### 2022-01-05 2426819
attempt to add global variables, see what happens
