### 2022-01-05 ed71115
add graphics and stats depends
