### 2022-01-05 cbaa7ad
update globals and add to NAMESPACE
