+++ NAMESPACE
@@ -101,5 +101,3 @@ importFrom(tidyr,expand)
 importFrom(tidyr,expand_grid)
 importFrom(tidyr,pivot_longer)
 importFrom(tidyr,pivot_wider)
-importFrom(graphics, legend, lines, text)
-importFrom(stats, median, quantile)
