### 2022-01-05 1370bc0
add more global variables; check if it breaks
