### 2022-01-05 aa55c5d
add more globals
