+++ R/globals.R
@@ -2,4 +2,4 @@
 # Based on suggestion from https://community.rstudio.com/t/how-to-solve-no-visible-binding-for-global-variable-note/28887
 
 utils::globalVariables(c("Code","S0","S1plus","AgeMat","nages","fmax","lambdaMax","K1plus","Type","dat","BycatchLvl","abund10","abund50","Depl",
-                         "prebuild50","abundrel50","depletion","fishing.rate","probability","S0.w","S1plus.w", "Attempt", "button_label","prebuild100","lo90","hi90","lo50","hi50","blvl","Range_lo","Range_hi","Value","param.mean","lh.params1","rel1p","yield"))
+                         "prebuild50","abundrel50","depletion","fishing.rate","probability","S0.w","S1plus.w", "Attempt", "button_label","prebuild100","lo90","hi90","lo50","hi50","blvl","Range_lo","Range_hi","Value","param.mean","lh.params1","rel1p","yield","variable","value","name","zero","Bycatch","abundrel10","abundrel20","AgeMat.w","nages.w","A.w","z.w","x","y","text","axis.no","high","low","name","sim","year","median","quantile","q50","q2.5","q97.5","q25","q75","Species","Parameter","value","Citation","name"))
