+++ R/18_multiplot_proj.R
@@ -22,65 +22,67 @@
 #' @return A plot of 50 percent and 90 percent confidence intervals of population projections if \code{spaghetti == FALSE} or a spaghetti plot with n individual projections if \code{spaghetti == n },  from \code{projections()}.
 #'
 #' @examples
-#' parms <- list(S0 = 0.944, S1plus = 0.99, 
-#' K1plus = 9000, 
-#' AgeMat = 18, nages = 20,  
-#' z = 2.39, lambdaMax = 1.02)
+#' parms <- list(
+#'   S0 = 0.944, S1plus = 0.99,
+#'   K1plus = 9000,
+#'   AgeMat = 18, nages = 20,
+#'   z = 2.39, lambdaMax = 1.02
+#' )
 #' InitDepl.vec <- c(0.1, 0.5, 0.9)
 #' BycatchCV <- 0.2
 #' nyears <- 100
-#' 
+#'
 #' high.list <- lapply(
-#' X = InitDepl.vec,
-#' function(x) {
-#'   projections(
-#'     NOut = 50,
-#'     ConstantRateBycatch = list(Rate = 0.3, CV = BycatchCV),
-#'     InitDepl = x,
-#'     lh.params = parms,
-#'     nyears = nyears,
-#'     obs_CV = 0.2
-#'   )
-#' }
+#'   X = InitDepl.vec,
+#'   function(x) {
+#'     projections(
+#'       NOut = 50,
+#'       ConstantRateBycatch = list(Rate = 0.3, CV = BycatchCV),
+#'       InitDepl = x,
+#'       lh.params = parms,
+#'       nyears = nyears,
+#'       obs_CV = 0.2
+#'     )
+#'   }
 #' )
-
+#'
 #' med.list <- lapply(
-#' X = InitDepl.vec,
-#' function(x) {
-#'   projections(
-#'     NOut = 50,
-#'     ConstantRateBycatch = list(Rate = 0.02, CV = BycatchCV),
-#'     InitDepl = x,
-#'     lh.params = parms,
-#'     nyears = nyears,
-#'     obs_CV = 0.2
-#'   )
-#' }
+#'   X = InitDepl.vec,
+#'   function(x) {
+#'     projections(
+#'       NOut = 50,
+#'       ConstantRateBycatch = list(Rate = 0.02, CV = BycatchCV),
+#'       InitDepl = x,
+#'       lh.params = parms,
+#'       nyears = nyears,
+#'       obs_CV = 0.2
+#'     )
+#'   }
 #' )
 #' low.list <- lapply(
-#' X = InitDepl.vec,
-#' function(x) {
-#'   projections(
-#'     NOut = 50,
-#'     ConstantRateBycatch = list(Rate = 0.001, CV = BycatchCV),
-#'     InitDepl = x,
-#'     lh.params = parms,
-#'     nyears = nyears,
-#'     obs_CV = 0.2
-#'   )
-#' }
+#'   X = InitDepl.vec,
+#'   function(x) {
+#'     projections(
+#'       NOut = 50,
+#'       ConstantRateBycatch = list(Rate = 0.001, CV = BycatchCV),
+#'       InitDepl = x,
+#'       lh.params = parms,
+#'       nyears = nyears,
+#'       obs_CV = 0.2
+#'     )
+#'   }
 #' )
 #' multiplot_proj(
-#' high.d1 = high.list[[1]], # d1 is the lowest depletion
-#' med.d1 = med.list[[1]],
-#' low.d1 = low.list[[1]],
-#' high.d2 = high.list[[2]],
-#' med.d2 = med.list[[2]],
-#' low.d2 = low.list[[2]],
[…]
