### 2022-01-05 e5a3d1e
more globals
