+++ NAMESPACE
@@ -1,5 +1,33 @@
 # Generated by roxygen2: do not edit by hand
 
+export(abund_rel)
+export(abund_rel_vec)
+export(add_trans)
+export(calc_z)
+export(ce)
+export(dynamics)
+export(find_msyr)
+export(get_diff)
+export(get_dz)
+export(get_f)
+export(get_mnpl)
+export(get_rf)
+export(getfecmax)
+export(ggradar)
+export(inv_logit)
+export(logit)
+export(make_ptable)
+export(multiplot_proj)
+export(npr)
+export(plot_bycatch_guesses)
+export(plot_pinnipeds)
+export(plot_proj)
+export(plot_yield_curve)
+export(pop_vs_yield)
+export(prob_rebuilt_goal)
+export(projections)
+export(rebuild_by_x)
+export(run_app)
 import(shiny)
 importFrom(config,get)
 importFrom(dplyr,bind_rows)
@@ -73,5 +101,3 @@ importFrom(tidyr,expand)
 importFrom(tidyr,expand_grid)
 importFrom(tidyr,pivot_longer)
 importFrom(tidyr,pivot_wider)
-importFrom("graphics", "legend", "lines", "text")
-importFrom("stats", "median", "quantile")
