### 2022-01-05 mcsiple/mmrefpoints#56 @mcsiple
Woo! Fixed with commit 3e9f8e4e . Just needed to add `mmrefpoints` to imports for `run_app()` fn.
