### 2022-01-06 44fb593
fix issue with kiteplots extending past 1 (sometimes go to 1.01*K) #54 and #57
