### 2022-01-06 baf8a3d
fix dataframe for multiple depletion levels
