### 2022-01-06 mcsiple/mmrefpoints#57 @mcsiple
Fixed multiplot issue with commit baf8a3d5
