### 2022-01-10 mcsiple/mmrefpoints#58 @tbrown122387
test coverage is low
Apologies if this echos some of the suggestions in #54, but the test coverage looks a bit low:

```
> covr::package_coverage("~/mmrefpoints/R/")
mmrefpoints Coverage: 31.21%
R/00_setup.R: 0.00%
R/01_npr.R: 0.00%
R/02_getfecmax.R: 0.00%
R/03_get_rf.R: 0.00%
R/05_inv_logit.R: 0.00%
R/06_get_f.R: 0.00%
R/07_ce.R: 0.00%
R/08_get_diff.R: 0.00%
R/09_find_msyr.R: 0.00%
R/10_get_mnpl.R: 0.00%
R/11_get_dz.R: 0.00%
R/12_calc_z.R: 0.00%
R/13_pop_vs_yield.R: 0.00%
R/15_plot_bycatch_guesses.R: 0.00%
R/16_add_trans.R: 0.00%
R/17_plot_proj.R: 0.00%
R/18_multiplot_proj.R: 0.00%
R/19_ggradar.R: 0.00%
R/20_dynamics.R: 0.00%
R/21_projections.R: 0.00%
R/23_prob_rebuilt_goal.R: 0.00%
R/24_abund_rel.R: 0.00%
R/25_abund_rel_vec.R: 0.00%
R/26_make_ptable.R: 0.00%
R/27_rebuild_by_x.R: 0.00%
R/golem_utils_server.R: 0.00%
R/golem_utils_ui.R: 0.00%
R/run_app.R: 0.00%
R/14_plot_yield_curve.R: 7.69%
R/app_config.R: 14.29%
R/app_server.R: 57.96%
R/22_plot_pinnipeds.R: 90.91%
R/04_logit.R: 100.00%
R/app_ui.R: 100.00%
```

openjournals/joss-reviews#3888
