### 2022-01-12 4a806f0
add labels with specific bycatch levels to figure legend for plot_proj() #54 #57
