### 2022-01-12 5bc3662
changed CV of catch to slider input instead of numeric - #54 and #57
