### 2022-01-18 01c1af2
cleanup ggradar()
