+++ R/19_ggradar.R
@@ -100,13 +100,7 @@ ggradar <- function(plot.data,
   # Declare required internal functions
 
   CalculateGroupPath <- function(df) {
-    # Converts variable values into a set of radial x-y coordinates
-    # Code adapted from a solution posted by Tony M to
-    # http://stackoverflow.com/questions/9614433/creating-radar-chart-a-k-a-star-plot-spider-plot-using-ggplot2-in-r
-    # Args:
-    #  df: Col 1 -  group ('unique' cluster / group ID of entity)
-    #      Col 2-n:  v1.value to vn.value - values (e.g. group/cluser mean or median) of variables v1 to v.n
-
+    
     path <- df[, 1]
 
     ## find increment
@@ -118,9 +112,6 @@ ggradar <- function(plot.data,
     for (i in levels(path)) {
       pathData <- subset(df, df[, 1] == i)
       for (j in c(2:ncol(df))) {
-        # pathData[,j]= pathData[,j]
-
-
         graphData <- rbind(graphData, data.frame(
           group = i,
           x = pathData[, j] * sin(angles[j - 1]),
@@ -167,7 +158,6 @@ ggradar <- function(plot.data,
     as.data.frame(axisData)
   }
   funcCircleCoords <- function(center = c(0, 0), r = 1, npoints = 100) {
-    # Adapted from Joran's response to http://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2
     tt <- seq(0, 2 * pi, length.out = npoints)
     xx <- center[1] + r * cos(tt)
     yy <- center[2] + r * sin(tt)
@@ -175,9 +165,6 @@ ggradar <- function(plot.data,
   }
 
   ### Convert supplied data into plottable format
-  # (a) add abs(centre.y) to supplied plot data
-  # [creates plot centroid of 0,0 for internal use, regardless of min. value of y
-  # in user-supplied data]
   plot.data.offset <- plot.data
   plot.data.offset[, 2:ncol(plot.data)] <- plot.data[, 2:ncol(plot.data)] + abs(centre.y)
 
@@ -226,8 +213,6 @@ ggradar <- function(plot.data,
     text = as.character(grid.mid)
   )
 
-  # Declare 'theme_clear', with or without a plot legend as required by user
-  # [default = no legend if only 1 group [path] being plotted]
   theme_clear <- ggplot2::theme_bw(base_size = 20) +
     ggplot2::theme(
       axis.text.y = ggplot2::element_blank(),
@@ -241,17 +226,6 @@ ggradar <- function(plot.data,
 
   if (plot.legend == FALSE) theme_clear <- theme_clear + ggplot2::theme(legend.position = "none")
 
-  # Base-layer = axis labels + plot extent
-  # [need to declare plot extent as well, since the axis labels don't always
-  # fit within the plot area automatically calculated by ggplot, even if all
-  # included in first plot; and in any case the strategy followed here is to first
-  # plot right-justified labels for axis labels to left of Y axis for x< (-x.centre.range)],
-  # then centred labels for axis labels almost immediately above/below x= 0
-  # [abs(x) < x.centre.range]; then left-justified axis labels to right of Y axis [x>0].
-  # This building up the plot in layers doesn't allow ggplot to correctly
-  # identify plot extent when plotting first (base) layer]
-
-  # base layer = axis labels for axes to left of central y-axis [x< -(x.centre.range)]
   base <- ggplot2::ggplot(axis$label) +
     ggplot2::xlab(NULL) +
     ggplot2::ylab(NULL) +
@@ -263,20 +237,20 @@ ggradar <- function(plot.data,
[…]
