### 2022-01-18 a82c26c
correct factor levels in ggradar example
