+++ man/ggradar.Rd
@@ -43,9 +43,14 @@ Since this code was originally written, ggradar has becomes its own standalone p
 \examples{
 plot.data <- data.frame(
   bycatch = factor(c(
-    "Lower end of bycatch range",
+    "Higher end of bycatch range",
     "Midpoint of bycatch range",
-    "Higher end of bycatch range"
+    "Lower end of bycatch range"
+  ),
+  ,levels = c(
+  "Higher end of bycatch range",
+  "Midpoint of bycatch range",
+  "Lower end of bycatch range"
   )),
   prebuild50 = c(0, 0, 1),
   prebuild100 = c(0, 0, 1),
@@ -59,6 +64,7 @@ ggradar(
   axis.label.size = 4,
   grid.label.size = 4
 )
+
 }
 \author{
 Ricardo Bion, modified slightly by Margaret Siple
