### 2022-01-18 901a79e
cleaned up ggradar, not sure why legend order is weird. It's taking forever to figure out so I'm moving onto the next thing for a bit.
