### 2022-01-18 89f8dec
update palette in example so it makes sense; remove manual.levels
