### 2022-01-19 084fd33
increase line width on geom_violin() sim outputs #54 #57
