### 2022-01-19 a9b12b5
add extract_df() function for turning outputs into a dataframe #54 #57
