### 2022-01-19 9f00cce
initial commit: turn trajectory list into df for easy output from shiny
