### 2022-01-19 efd45df
add documentation for functions to make outputs into an easy txt table #54 #57
