+++ man/extract_df.Rd
@@ -1,0 +1,22 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/28_extract_df.R
+\name{extract_df}
+\alias{extract_df}
+\title{Extract results from `dynamics()` function as a dataframe}
+\usage{
+extract_df(x)
+}
+\arguments{
+\item{x}{Outputs from a call to dynamics(), which are a named list}
+}
+\value{
+a dataframe with population trajectories, initial depletion, and bycatch levels that can be printed as a table.
+}
+\description{
+#unnest results from dynamics() and turn them into a dataframe that you can bind together
+}
+\examples{
+x <- dynamics(S0 = 0.944, S1plus = 0.99, K1plus = 9000, AgeMat = 17, InitDepl = 0.6, ConstantCatch = NA, ConstantF = rep(0.01, times = 100), z = 2.39, nyears = 100, nages = 25, lambdaMax = 1.04)
+extract_df(x)
+
+}
