+++ NAMESPACE
@@ -6,6 +6,7 @@ export(add_trans)
 export(calc_z)
 export(ce)
 export(dynamics)
+export(extract_df)
 export(find_msyr)
 export(get_diff)
 export(get_dz)
@@ -28,6 +29,7 @@ export(prob_rebuilt_goal)
 export(projections)
 export(rebuild_by_x)
 export(run_app)
+export(traj_list_to_df)
 import(shiny)
 importFrom(config,get)
 importFrom(dplyr,bind_rows)
