### 2022-01-25 d62285c
add option to download csv file of all projections #54 #57
