### 2022-01-26 40a382f
round digits in traj for raw outputs
