### 2022-01-26 bcc7db0
update publications that cite this pkg
