### 2022-01-26 mcsiple/mmrefpoints#54 @mcsiple
Updated citations for papers that cite this tool with commit bcc79b09
