### 2022-01-26 5bae34b
add table to Solver tab with params from Advanced tab #54 #57
