+++ R/app_server.R
@@ -1,6 +1,6 @@
 #' The application server-side
-#' 
-#' @param input,output,session Internal parameters for {shiny}. 
+#'
+#' @param input,output,session Internal parameters for {shiny}.
 #'     DO NOT REMOVE.
 #' @import shiny
 #' @importFrom dplyr select mutate rename filter rename_with recode_factor funs
@@ -9,106 +9,107 @@
 #' @importFrom magrittr %>%
 #' @importFrom ggplot2 ggplot geom_violin geom_point theme_classic theme xlab ylab scale_fill_manual scale_colour_viridis_c theme_minimal theme geom_hline
 #' @noRd
-app_server <- function( input, output, session ) {
+app_server <- function(input, output, session) {
   # List the first level callModules here
-  
+
   # TRANSLATOR --------------------------------------------------------------
-  
+
   observeEvent(input$selected_language, {
     print(paste("Language change:", input$selected_language))
     shiny.i18n::update_lang(session, input$selected_language)
   })
-  
+
   # HOMEPAGE ----------------------------------------------------------------
   output$splash <- renderUI({
     htmltools::HTML(
       switch(input$selected_language,
-                           "en" = splash_en,
-                           "es" = splash_es,
-                           "fr" = splash_fr
-    ))
+        "en" = splash_en,
+        "es" = splash_es,
+        "fr" = splash_fr
+      )
+    )
   })
-  
+
   output$tn_simple <- renderUI({
     htmltools::HTML(switch(input$selected_language,
-                           "en" = tn_en,
-                           "es" = tn_es,
-                           "fr" = tn_fr
+      "en" = tn_en,
+      "es" = tn_es,
+      "fr" = tn_fr
     ))
   })
-  
+
   output$tn_advanced <- renderUI({
     htmltools::HTML(switch(input$selected_language,
-                           "en" = tn_en_adv,
-                           "es" = tn_es_adv,
-                           "fr" = tn_fr_adv
+      "en" = tn_en_adv,
+      "es" = tn_es_adv,
+      "fr" = tn_fr_adv
     ))
   })
-  
+
   observeEvent(input$jumpToSimple, {
     updateTabsetPanel(session, "inTabset",
-                      selected = "simple"
+      selected = "simple"
     )
   })
-  
+
   observeEvent(input$jumpToAdv, {
     updateTabsetPanel(session, "inTabset",
-                      selected = "advanced"
+      selected = "advanced"
     )
   })
-  
-  
+
+
   # SIMPLE ------------------------------------------------------------------
   # translate radio buttons and dropdown menus
   observe({
     updateSelectInput(session,
-                      inputId = "type_simple",
-                      label = i18n$t("Mammal life history type (choose one):"),
-                      choices = stats::setNames(paste0(mmrefpoints::lh$Code, "_simple"), i18n$t(mmrefpoints::lh$Type))
+      inputId = "type_simple",
+      label = i18n$t("Mammal life history type (choose one):"),
+      choices = stats::setNames(paste0(mmrefpoints::lh$Code, "_simple"), i18n$t(mmrefpoints::lh$Type))
     ) # / update SelectInput
   })
-  
+
   observeEvent(input$type, {
     x <- input$type
     updateNumericInput(session, "global.S0", value = c(lh.table()[paste(x), "S0"]))
     updateNumericInput(session, "global.S1plus", value = c(lh.table()[paste(x), "S1plus"]))
     updateNumericInput(session, "global.AgeMat", value = c(lh.table()[paste(x), "AgeMat"]))
   })
-  
+
   observeEvent(input$popsize_usr_simple, {
     updateSliderInput(session, "constantcatch_simple",
-                      value = c(50, 100),
-                      min = 0,
-                      max = input$popsize_usr_simple
[…]
