### 2022-01-27 bf87157
Change params in dynamics() into a list (now matches projections())
