### 2022-01-27 mcsiple/mmrefpoints#57 @mcsiple
I have addressed all but two of the above:
1) The bycatch factor levels in the kite plot. The `ggradar()` function makes this weirdly difficult and I haven't figured out how to fix it. Since the legend colors are correctly assigned to bycatch levels I am leaving this for now, unless Hao feels strongly about it.
2) The side by side design is a little weird before the projections are run, but it does allow the user to see the parameters they entered and the projections at the same time. I am open to others' opinion on this, but I am inclined to keep the sidebar-style layout. Maybe @tessabfrancis can help with some wisdom/opinion.

Everything else is incorporated now and commits have been tagged both here and in Hao Ye's original review issue, #54 .
