### 2022-01-27 0bafcc9
Fix author list and version in CITATION.cff #54
