### 2022-01-27 8dc3c87
update badge and call to dynamics()
