+++ README.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 # mmrefpoints: Projecting long-term marine mammal abundance with bycatch
 
 <!-- badges: start -->
-[![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://lifecycle.r-lib.org/articles/stages.html#experimental)
+[![Lifecycle: experimental](https://img.shields.io/badge/lifecycle-experimental-orange.svg)](https://github.com/mcsiple/mmrefpoints/issues)
 [![DOI](https://zenodo.org/badge/344858328.svg)](https://zenodo.org/badge/latestdoi/344858328)
 <!-- badges: end -->
 
@@ -100,13 +100,12 @@ knitr::kable(x)
 
 To create a single projection for a marine mammal population, use the `dynamics()` function:
 ```{r message=FALSE}
-x <- mmrefpoints::dynamics(S0 = 0.944, S1plus = 0.99, 
-                           K1plus = 9000, AgeMat = 17,
+x <- mmrefpoints::dynamics(lh.params = list(S0 = 0.944, S1plus = 0.99, 
+                           K1plus = 9000, AgeMat = 17, z = 2.39, nages = 25, lambdaMax = 1.04),
                            InitDepl = 0.6, 
                            ConstantCatch = NA, 
                            ConstantF = rep(0.01, times = 100), 
-                           z = 2.39, nyears = 100, 
-                           nages = 25, lambdaMax = 1.04)
+                           nyears = 100)
 plot(1:100, x$TotalPop, type = 'l', xlab = "Year", ylab = "Population size")
 ```
 
