### 2022-01-27 mcsiple/mmrefpoints#54 @mcsiple
Hi Hao,

I used issue #57 to keep track of all the functionality reviews (thank you for those) and have, I think, addressed all of them. I'm going to address the others as I go, here. I will try to minimize the number of separate comments so you don't get a ton of notifications. If you don't mind, for my own record-keeping and issue tracking purposes, would you mind checking off the sections that you feel are properly addressed? You could do it here or in the [JOSS review](https://github.com/openjournals/joss-reviews/issues/3888) or both.

## General checks
- I have fixed the author list in CITATION.cff
- I would love to turn on auto-archiving for this repo and am going to attempt it once I think I have addressed all of your comments and @tbrown122387 's, to see if it added the other authors. I remember having trouble earlier adding authors to the Zenodo doi.

## Functionality
- Fixed `disable_autoload` issue
- See #57 for revisions to functionality. The only things I didn't incorporate were the order of the legend colors in the kite plot (I spent a couple hours trying to fix the order but have abandoned the attempt; I don't think it affects usability even though it's a little weird) and the side bar layout (we agreed on this layout with the UI folks we consulted, and it is intended to keep the projections plot alongside the input parameters so the users can see the relationship between them)

## Documentation
### Example usage
- I agree that a Markdown report could be helpful. For now, I am adding this as a feature request and if we get support for it from stakeholders I will add it. (#59 )

### Functionality documentation
- I will look into whether I can easily generate the model description vignette from the About the Model tab contents. It shouldn't be too hard but I haven't done it yet.
- I added the code version of the parameters to the table in "About the Model" -- if this is adequate, I will do the same to the other languages and to the vignette.

### Automated tests
I still need to work on these. Overlap with @tbrown122387 's comment and definitely something that I need to work on. There are Shiny-specific tests but I think the most important ones are for the embedded functions for population dynamics, etc.

### Community guidelines
- Badge link in README has been fixed (now points to Issues, thank you) 8dc3c875
- I think a CONTRIBUTING.md file is a great idea and will work on it, but haven't created it yet. Added as issue #60

## Software paper
### State of the field
- rSPAMM has been added to the paper -- good find.
- References: updated references in `paper.md`
