### 2022-01-27 9550638
add inv logit test
