### 2022-01-27 b711532
add tests for setup
