### 2022-01-27 4c00247
add tests for npr()
