### 2022-01-27 29ef2c5
Elaborate on getfecmax example and add test #58
