### 2022-01-27 3a78922
add tests for get_rf; slight mod to example code
