### 2022-01-27 714f06c
cleanup get_rf
