### 2022-01-27 f50b4e2
Add test for get_f() #58 - experiment with tagging JOSS review -- openjournals/joss-reviews#3888
