### 2022-01-27 #3888 @mcsiple
Thank you all for your patience and both Hao and Taylor for your reviews. I have been working on addressing both reviewers' comments in the following issues:

1. @ha0ye 's review notes:  mcsiple/mmrefpoints#57
2. Detailed responses and commits related to functionality review from @ha0ye : mcsiple/mmrefpoints#54
3. @tbrown122387 's coverage comments: mcsiple/mmrefpoints#58

I have been working through the testing infrastructure and should be done soon. I have now made modifications addressing everything (I think) except test coverage (should be done soon) and the version updating issue (will make a new release when I'm done writing tests; zenodo is now auto-linked to the github repo so it should automatically update authors and version # when I push changes).

In the first version I released, I accidentally only included the first three authors; if anyone has experience changing the author list for a zenodo-linked repo, I would love your guidance. Ideally the next release (0.1.2) will have all the authors that are in CITATION.cff.

Thank you both again, and more coming soon.
