### 2022-01-27 4ba4226
add another test for get_f() #58
