### 2022-01-27 9b2e0d6
add tests for ce() #58 openjournals/joss-reviews#3888
