### 2022-01-27 98713aa
cleanup ce() test style
