### 2022-01-27 a2e8ded
add test for get_diff() #58 openjournals/joss-reviews#3888
