### 2022-01-27 07175b9
move getfecmax() test to correct dir (oops)
