+++ R/09_find_msyr.R
@@ -4,7 +4,7 @@
 #'
 #' @param E.start a starting guess for bycatch mortality rate that will result in FMSY (numeric value)
 #' @param lh.params a list of life history parameters (S0, S1plus, nages, AgeMat, lmabdaMax, K1plus, and z)
-#' @param fmax Max theoretical fecundity (num)
+#' @param fmax Max theoretical fecundity (numeric)
 #' 
 #' @examples
 #' # Set parameters
