### 2022-01-27 731855a
add test for find_msyr() #58 openjournals/joss-reviews#3888
