### 2022-01-27 #3888 @ha0ye
> In the first version I released, I accidentally only included the first three authors; if anyone has experience changing the author list for a zenodo-linked repo, I would love your guidance. Ideally the next release (0.1.2) will have all the authors that are in CITATION.cff.

I believe you can go in to Zenodo -> Upload to manually select and then update the info on an existing deposit. Or if you're set up for auto-archiving of releases, then the newer DOI will have the updated information.
