### 2022-01-29 4088c77
cleanup examples
