+++ R/12_calc_z.R
@@ -9,10 +9,10 @@
 #' @export
 #'
 #' @examples
-#' test.z <- calc_z(MNPL_in = 0.5,
+#' calc_z(MNPL_in = 0.5,
 #' lh.params_in = list(S0 = 0.944, S1plus = 0.99, AgeMat = 17, nages = 19,
-#' fmax = 0.29, z = 2.39, lambdaMax = 1.04, K1plus = 9000))
-#' test.z
+#' fmax = 0.29, lambdaMax = 1.04, K1plus = 9000))
+#' 
 calc_z <- function(MNPL_in, lh.params_in) {
   # Checks
   if(MNPL_in < 0 | MNPL_in > 1){
