### 2022-01-29 31d038f
cleanup examples
