### 2022-01-29 13e25b5
add test for get_mnpl and calc_z #58 openjournals/joss-reviews#3888
