### 2022-01-29 d722aa2
add test for plotting yield curve #58 openjournals/joss-reviews#3888
