### 2022-01-29 f373125
add tests for plot_bycatch_guesses() #58 openjournals/joss-reviews#3888
