+++ NAMESPACE
@@ -103,3 +103,4 @@ importFrom(tidyr,expand)
 importFrom(tidyr,expand_grid)
 importFrom(tidyr,pivot_longer)
 importFrom(tidyr,pivot_wider)
+importFrom(utils,write.csv)
