+++ man/dynamics.Rd
@@ -35,7 +35,9 @@ The population model is a single-sex age-structured model in which the number of
 }
 \examples{
 # Generate a time series of abundance for a bowhead whale
-dynamics(lh.params = list(S0 = 0.944, S1plus = 0.99, K1plus = 9000, AgeMat = 17,nages = 25,z = 2.39, lambdaMax = 1.04),
+dynamics(lh.params = list(S0 = 0.944, S1plus = 0.99, 
+K1plus = 9000, AgeMat = 17,nages = 25,
+z = 2.39, lambdaMax = 1.04),
  InitDepl = 0.6, ConstantCatch = NA, ConstantF = rep(0.01, times = 100), 
  nyears = 100)
 }
