### 2022-01-29 544302b
add test for dynamics #58 openjournals/joss-reviews#3888
