+++ CITATION.cff
@@ -33,7 +33,7 @@ authors:
     orcid: https://orcid.org/0000-0001-9390-6693
   - family-names: Víkingsson
     given-names: Gísli
-    orcid: https://orcid.org/00000-0002-4501-193X
+    orcid: https://orcid.org/0000-0002-4501-193X
   - family-names: Wade
     given-names: Paul R.
     orcid: https://orcid.org/0000-0003-2428-9323
