### 2022-01-29 5ec150a
add tests for add_trans() and make_ptable #58 openjournals/joss-reviews#3888
