### 2022-01-29 3d0f06d
add tests for plotting functions #58 openjournals/joss-reviews#3888
