### 2022-01-29 8d90bc5
add test for rebuild_by_x()
