### 2022-01-29 e01c4b7
add tests for extract_df()
