### 2022-01-30 fa28224
add test for traj_list_to_df() #58 openjournals/joss-reviews#3888
