### 2022-01-30 3d99ac8
add test for run_app()
