### 2022-01-30 7186b16
add another test to rebuild_by_x()
