+++ man/rebuild_by_x.Rd
@@ -15,25 +15,25 @@ rebuild_by_x(
 )
 }
 \arguments{
-\item{needf.start}{starting guess for the bycatch mortality rate needed to recover the population}
+\item{needf.start}{Starting guess for the bycatch mortality rate needed to recover the population}
 
-\item{init.depl.w}{initial depletion (a fraction)}
+\item{init.depl.w}{Initial depletion (a fraction)}
 
 \item{goal.w}{Population size goal (number of whales or pinnipeds) to rebuild to, expressed as a whole number}
 
-\item{desired.prob.w}{what probability you want (e.g., 0.75 probability that the population will rebuild to X in Y years)}
+\item{desired.prob.w}{What probability you want (e.g., 0.75 probability that the population will rebuild to X in Y years)}
 
-\item{when.w}{the year Y when rebuilding is desired by (number of years into future; current year = 0)}
+\item{when.w}{The year Y when rebuilding is desired by (number of years into future; current year = 0)}
 
-\item{lh.params.w}{a list of life history parameters used by \code{projections()}}
+\item{lh.params.w}{A list of life history parameters used by \code{projections()}}
 
-\item{fixed.cv.catch.w}{the CV of the bycatch rate - should be fixed}
+\item{fixed.cv.catch.w}{The CV of the bycatch rate - should be fixed}
 }
 \value{
 The bycatch rate that will result in the specified rebuilding goal. If Shiny is running, it will return a list containing the bycatch rate \code{f} and a matrix of guesses that \code{optim()} has searched through to find the solution.
 }
 \description{
-takes a performance goal (what level you want to rebuild to) and a time window (how long you want that to take) and calculates what the bycatch rate needs to be
+Takes a performance goal (what level you want to rebuild to) and a time window (how long you want that to take) and calculates what the bycatch rate needs to be
 }
 \examples{
 rebuild_by_x(
