### 2022-01-30 #3888 @mcsiple
Alright! I believe I am done making revisions. A few notes:

## Author list in CITATION.cff and on Zenodo
I was able to properly update the `.zenodo.json` file so that it automatically updates the [Zenodo entry](https://zenodo.org/record/5919455#.YfYZzVjMI-Q) when I make a new release on GitHub. [Release v0.1.6](https://github.com/mcsiple/mmrefpoints/releases/tag/v.0.1.6) has all the changes incorporated. I think it is fully ready for the reviewers to look over again.

## Test coverage
Test coverage is addressed by the commits above and in mcsiple/mmrefpoints/issues/58. The coverage is not 100%; this is partly because some of the functions that `covr::report()` finds are `golem` functions that I did not write.

## Not sure if I need to tag whedon
Tagging @ha0ye , @tbrown122387 , and @fboehm  here to let you know I have finished revisions.
