### 2022-01-30 mcsiple/mmrefpoints#58 @tbrown122387
niceee!

```
> covr::package_coverage("~/mmrefpoints/R/")
mmrefpoints Coverage: 65.11%
R/golem_utils_server.R: 0.00%
R/golem_utils_ui.R: 0.00%
R/run_app.R: 0.00%
R/27_rebuild_by_x.R: 11.11%
R/app_config.R: 14.29%
R/00_setup.R: 19.05%
R/02_getfecmax.R: 50.00%
R/app_server.R: 56.93%
R/14_plot_yield_curve.R: 61.54%
R/17_plot_proj.R: 66.43%
R/18_multiplot_proj.R: 67.07%
R/16_add_trans.R: 70.00%
R/03_get_rf.R: 72.73%
R/12_calc_z.R: 76.92%
R/13_pop_vs_yield.R: 77.55%
R/23_prob_rebuilt_goal.R: 80.00%
R/06_get_f.R: 81.08%
R/24_abund_rel.R: 81.82%
R/01_npr.R: 83.33%
R/25_abund_rel_vec.R: 85.71%
R/07_ce.R: 87.50%
R/09_find_msyr.R: 88.89%
R/08_get_diff.R: 89.66%
R/21_projections.R: 90.12%
R/26_make_ptable.R: 90.48%
R/22_plot_pinnipeds.R: 90.91%
R/15_plot_bycatch_guesses.R: 92.73%
R/20_dynamics.R: 93.98%
R/19_ggradar.R: 94.71%
R/04_logit.R: 100.00%
R/05_inv_logit.R: 100.00%
R/10_get_mnpl.R: 100.00%
R/11_get_dz.R: 100.00%
R/28_extract_df.R: 100.00%
R/28_traj_list_to_df.R: 100.00%
R/app_ui.R: 100.00%
```
