### 2022-02-02 41eaa8d
update in-text citations for other OMF projects in paper.md
