+++ paper.bib
@@ -186,15 +186,35 @@
 	year = {2019},
 }
 
-@article{wade_best_nodate,
+@article{wade_best_2021,
 	title = {Best {Practices} for {Assessing} and {Managing} {Bycatch} of {Marine} {Mammals}},
-	abstract = {Bycatch in marine fisheries is one of the leading sources of human-caused mortality for marine mammals, and has contributed to substantial declines of many marine mammal populations and species, and even the extinction of at least one. Schemes for evaluating marine mammal bycatch largely rely on estimates of abundance and bycatch, which are needed for calculating biological reference points and determining conservation status. However, obtaining these estimates is resource intensive and takes careful long-term planning. The need for assessments of marine mammal bycatch in fisheries is expected to increase worldwide due to the recently implemented seafood import provisions of the USA Marine Mammal Protection Act. Thus, managers and other stakeholders need reliable, standardized methods for collecting data to estimate abundance and bycatch rates. In some cases, managers will be starting with little or no data and no system in place to collect data. We outline a comprehensive framework for managing bycatch of marine mammals. We describe and provide guidance on (1) planning for an assessment of bycatch, (2) collecting appropriate data (e.g., abundance and bycatch estimates), (3) assessing bycatch and calculating reference points, and (4) using the results of the assessment to guide marine mammal bycatch reduction. We provide a brief overview of available mitigation techniques to reduce marine mammal bycatch in various fisheries. This paper provides information for scientists and resource managers in the hope that it will lead to new or improved programs for assessing marine mammal bycatch, establishing best practices, and enhancing marine mammal conservation globally.},
-	author = {Wade, P. R. and Long, K. J. and Francis, T. B. and Punt, A. E. and Hammond, P. S. and Heinemann, D. and Moore, J. E. and Reeves, R. R. and Sepulveda, M. and Sullaway, G. and Sigurðsson, G. M. and Siple, M. C. and Vikingsson, G. and Williams, R. and Zerbini, A. N.},
+	volume = {8},
+	issn = {2296-7745},
+	url = {https://www.frontiersin.org/article/10.3389/fmars.2021.757330},
+	doi = {10.3389/fmars.2021.757330},
[…]
