+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,17 +1,22 @@
 ---
 name: Bug report
 about: Create a report to help us improve
-title: ''
-labels: ''
+title: '[BUG]'
+labels: 'bug'
 assignees: ''
 
 ---
 
 **Describe the bug**
-A clear and concise description of what the bug is.
+A clear and concise description of what the bug is. Please check whether you encountered this bug:
+
+- [ ] while using the Shiny app in the browser (https://msiple.shinyapps.io/mmrefpoints/)
+- [ ] while using the Shiny app locally (i.e., using `run_app()` in R on your computer)
+- [ ] while using the `mmrefpoints` functions locally
 
 **To Reproduce**
 Steps to reproduce the behavior:
+
 1. Go to '...'
 2. Click on '....'
 3. Scroll down to '....'
@@ -24,15 +29,16 @@ A clear and concise description of what you expected to happen.
 If applicable, add screenshots to help explain your problem.
 
 **Desktop (please complete the following information):**
+
  - OS: [e.g. iOS]
  - Browser [e.g. chrome, safari]
  - Version [e.g. 22]
 
-**Smartphone (please complete the following information):**
- - Device: [e.g. iPhone6]
- - OS: [e.g. iOS8.1]
- - Browser [e.g. stock browser, safari]
- - Version [e.g. 22]
+**Information about your version of R**
+
+ - R version (e.g., 4.1.2, get this by running `R.Version()` in the console)
+ - Session info (`sessionInfo()` in the console)
+ - How you are accessing R (RStudio and version, R GUI)
 
 **Additional context**
 Add any other context about the problem here.
