+++ CONTRIBUTING.md
@@ -1,0 +1,115 @@
+# Contributing
+
+## Welcome
+Thank you for your interest in contributing to `mmrefpoints`! We hope to maintain this package as a resource for a broad set of stakeholders and modelers.
+
+## Table of Contents
+
+[Short Links](#short-links)
+
+[How Can I Contribute?](#how-can-i-contribute)
+
+[Style Guides](#style-guides)
+
+
+## Short Links
+* View the rest of the Ocean Modeling Forum Marine Mammal Bycatch Working Group products [here](https://oceanmodelingforum.org/working-groups/marine-mammal-bycatch-working-group/)
+* Report bugs [here](https://github.com/mcsiple/mmrefpoints/issues)
+* Contact Margaret at margaret (dot) siple (at) noaa.gov
+
+## How Can I Contribute?
+ * [Reporting Bugs](#reporting-bugs)
+ * [Suggesting Enhancements](#suggesting-enhancements)
+ * [Pull Requests](#pull-requests)
+ 
+## Style Guides
+* [Write an issue on GitHub](#issue-style-guides)
+* [Code style guides](#code-style-guides)
+
+## Reporting Bugs 
+Bugs are tracked as [GitHub issues](https://guides.github.com/features/issues/). If you find a bug, please make your report as detailed as possible. Fill out the [required bug report template](https://github.com/mcsiple/mmrefpoints/blob/master/.github/ISSUE_TEMPLATE/bug_report.md); the information it asks for helps us resolve issues faster.
+
+**Note:** If you find a **Closed** issue that looks like it is the same thing that you're experiencing, open a new issue and include a link to the original issue in the body of your new one (you can tag issues by a hashtag (#) followed by the issue number).
+
+## Suggesting Enhancements 
+:pencil: We are always interested in how to make this package more useful and accessible. If you have an enhancement to suggest, please submit a [Feature Request](https://github.com/mcsiple/mmrefpoints/blob/master/.github/ISSUE_TEMPLATE/feature_request.md). 
+
+### How do I submit a good enhancement suggestion?
+Enhancement suggestions are tracked as [GitHub issues](https://guides.github.com/features/issues/). Create an issue on the mmrefpoints repository and include the following:
+
+* **Use a clear and descriptive title** for the issue to identify the suggestion.
+* **Provide a step-by-step description of the suggested enhancement** in as many details as possible.
+* **Provide specific examples to demonstrate the steps**. Include copy/pasteable snippets which you use in those examples, as [Markdown code blocks](https://help.github.com/articles/markdown-basics/#multiple-lines).
+* **Describe the current behavior** and **explain which behavior you expected to see instead** and why.
+* **Include screenshots and animated GIFs** which help you demonstrate the steps or point out the part of Atom which the suggestion is related to. You can use [this tool](https://www.cockos.com/licecap/) to record GIFs on macOS and Windows, and [this tool](https://github.com/colinkeenan/silentcast) or [this tool](https://github.com/GNOME/byzanz) on Linux.
+* **Explain why this enhancement would be useful** to most mmrefpoints users and stakeholders.
+* **Specify what version of mmrefpoints you are using** and whether you are using the Shiny app in the browser, the local Shiny app within the mmrefpoints package, or the functions in mmrefpoints by themselves.
+* **Specify the name and version of the OS you're using.**
+
+
+## Pull requests
+If you have a code suggestion in hand, please submit a pull request. 
+
+## Write an issue on GitHub
+### Git commit messages
+(many of these are lifted from the Atom [style guide](https://github.com/atom/atom/blob/master/CONTRIBUTING.md) for commit messages)
+
+* Use the present tense ("Add feature" not "Added feature")
[…]
