### 2022-02-02 d12c8cd
update CITATION.cff and DESCRIPTION
