+++ DESCRIPTION
@@ -1,40 +1,42 @@
 Package: mmrefpoints
+Type: Package
 Title: Project Marine Mammal Populations and Calculate Reference Points
 Version: 0.1.6
-Authors@R: c(person(c("Margaret"","C."), "Siple", 
-email = "margaret.siple@noaa.gov", 
-role = c("cre","aut"), 
-comment = c(ORCID = "0000-0002-4260-9856")),
-person(c("André", "E."), "Punt", 
-role = "aut", comment = c(ORCID = "0000-0001-8489-2488")),
-person(c("Tessa", "B."), "Francis", 
-role = "aut", comment = c(ORCID = "0000-0002-3383-5392")),
-person(c("Philip", "S."), "Hammond", 
-role = "aut", comment = c(ORCID = "0000-0002-2381-8302")),
-person("Dennis", "Heinemann", 
-role = "aut", comment = c(ORCID = "0000-0002-1434-2445")),
-person(c("Kristy", "J."), "Long", 
-role = "aut", comment = c(ORCID = "0000-0001-6970-0935")),
-person(c("Jeffrey", "E."), "Moore", 
-role = "aut", comment = c(ORCID = "0000-0003-3715-7442")),
-person(c("Randall", "R."), "Reeves", 
-role = "aut", comment = c(ORCID = "0000-0002-6512-6507")),
-person("Sepúlveda", "Maritza", 
-role = "aut", comment = c(ORCID = "0000-0002-1403-176X")),
-person(c("Guðjón","Már"), "Sigurðsson", 
-role = "aut", comment = c(ORCID = "0000-0001-9390-6693")),
-person("Gísli", "Víkingsson", 
-role = "aut", comment = c(ORCID = "0000-0002-4501-193X")),
-person(c("Paul", "R."), "Wade", 
-role = "aut", comment = c(ORCID = "0000-0003-2428-9323")),
-person("Rob", "Williams", 
-role = "aut", comment = c(ORCID = "0000-0001-7496-453X")),
-person(c("Alexandre", "N."), "Zerbini", 
-role = "aut", comment = c(ORCID = "0000-0002-9776-6605")))
+Authors@R: c(person(c("Margaret","C."), "Siple", 
+  email = "margaret.siple@noaa.gov", 
+  role = c("cre","aut"), 
+  comment = c(ORCID = "0000-0002-4260-9856")),
+  person(c("André", "E."), "Punt", 
+  role = "aut", comment = c(ORCID = "0000-0001-8489-2488")),
+  person(c("Tessa", "B."), "Francis", 
+  role = "aut", comment = c(ORCID = "0000-0002-3383-5392")),
+  person(c("Philip", "S."), "Hammond", 
+  role = "aut", comment = c(ORCID = "0000-0002-2381-8302")),
+  person("Dennis", "Heinemann", 
+  role = "aut", comment = c(ORCID = "0000-0002-1434-2445")),
+  person(c("Kristy", "J."), "Long", 
+  role = "aut", comment = c(ORCID = "0000-0001-6970-0935")),
+  person(c("Jeffrey", "E."), "Moore", 
+  role = "aut", comment = c(ORCID = "0000-0003-3715-7442")),
+  person(c("Randall", "R."), "Reeves", 
+  role = "aut", comment = c(ORCID = "0000-0002-6512-6507")),
+  person("Sepúlveda", "Maritza", 
+  role = "aut", comment = c(ORCID = "0000-0002-1403-176X")),
+  person(c("Guðjón","Már"), "Sigurðsson", 
+  role = "aut", comment = c(ORCID = "0000-0001-9390-6693")),
+  person("Gísli", "Víkingsson", 
[…]
