### 2022-02-02 d020f0e
update versions and paper.md for R1 release
