+++ CITATION.cff
@@ -45,5 +45,5 @@ authors:
     orcid: https://orcid.org/0000-0002-9776-6605
     
 title: mcsiple/mmrefpoints: Journal submission pre-release
-version: v.0.1.6
-date-released: 2022-1-29
+version: v.0.1.7
+date-released: 2022-2-2
