### 2022-02-04 207b008
update README links to contribution guide and link to permanent doi on Zenodo
