+++ README.Rmd
@@ -68,9 +68,13 @@ devtools::install_github("mcsiple/mmrefpoints")
 
 ## Contributing [![contributions welcome](https://img.shields.io/badge/contributions-welcome-brightgreen.svg?style=flat)](https://github.com/dwyl/esta/issues)
 ### Community guidelines
-We would like this package to be sustainable in the long term and welcome contributions. If you encounter a bug, please leave a note on the [Issues page](https://github.com/mcsiple/mmrefpoints/issues). You can also leave comments there about additional functionality (please add the "enhancement" label to your issue). 
+We would like this package to be sustainable in the long term and welcome contributions. If you are interested in contributing, please check out our [Contribution Guide](https://github.com/mcsiple/mmrefpoints/blob/master/CONTRIBUTING.md). 
 
-If you are interested in contributing, we direct you to the R package [contribution advice](http://r-pkgs.had.co.nz/git.html) from Hadley Wickham.
+Bugs and enhancements are tracked through GitHub issues. If you have a bug to report, there is a [bug report template](https://github.com/mcsiple/mmrefpoints/blob/master/.github/ISSUE_TEMPLATE/bug_report.md) to help maximize the benefit of your report for everyone. The same goes for requests for [enhancements](https://github.com/mcsiple/mmrefpoints/blob/master/.github/ISSUE_TEMPLATE/feature_request.md). 
+
+
+
+
 
 ## Accessing the MMBIET Shiny app
 The functions in this package can also be accessed through the Shiny app for this project, which is located online [here](https://msiple.shinyapps.io/mmrefpoints/). The app provides an easy way to explore outcomes and print out a report with inputs and outputs. 
@@ -180,4 +184,6 @@ To cite this package or the MMBIET Shiny app, please use the following citation:
 > Margaret C. Siple, André E. Punt, Tessa B. Francis, Philip S. Hammond, Dennis Heinemann, Kristy J. Long, Jeffrey E. Moore,
   Randall R. Reeves, Sepúlveda Maritza, Guðjón Már Sigurðsson, Gísli Víkingsson, Paul R. Wade, Rob Williams and Alexandre N.
   Zerbini (NA). mmrefpoints: Project Marine Mammal Populations and Calculate Reference Points. R package version 0.1.6.
-  url: <https://github.com/mcsiple/mmrefpoints> doi: 10.5281/zenodo.4758402
+  url: <https://github.com/mcsiple/mmrefpoints> doi: 10.5281/zenodo.4758401
+  
+NOTE that if you want to cite all versions of the software, you can use the doi [10.5281/zenodo.4758401](https://zenodo.org/record/5949332#.Yf1infXMI-Q). When additional releases happen, there will be a doi for each new release as well. 
