### 2022-03-03 mcsiple/mmrefpoints#54 @ha0ye
Yep, sorry. Had a few more functional notes, which I edited into the original post above.
