### 2022-03-03 3cd3c51
fix desc() issue and duplicated LazyData line for #54
