### 2022-03-03 mcsiple/mmrefpoints#54 @mcsiple
OK -- just fixed those two first functionality issues (thanks!). Commit noted above. As for the third bullet point, the parameters do show up in the "Solve for Bycatch" tab when the plot and solution are rendered. The final result looks like this:

That's what you meant for that one, right? LMK if there is anything else and thank you for all the time you've put into testing and reviewing!
