### 2022-03-03 mcsiple/mmrefpoints#54 @mcsiple
Ah, I see what you mean. I added that little table to the PBR tab and put in a translation for it. Thanks! I'm going to close this issue now if that's cool with you!
