### 2022-03-03 #3888 @mcsiple
Thanks @ha0ye -- and thank you for adding the Markdown child doc thing as its own issue. I  don't know if this is required, but the automated review thingy might be waiting until you check off boxes above.

@fboehm Let me know if there are other things I should address, otherwise I eagerly await your decision!
