### 2022-03-06 #3888 @fboehm
@mcsiple - the reviewers have recommended your submission for publication. Before we officially publish it, please make the following changes to your paper.md and bib file:

There's one small typo that needs fixing:

line 56 (of the pdf): delete "based"

Also, please cite R and shiny and include the references in your bibliography.

We can proceed towards publication once these changes are made.
