### 2022-03-06 e885e73
fix typo in paper.md and add citations for Shiny and R for https://github.com/openjournals/joss-reviews/issues/3888#
