+++ paper.md
@@ -85,7 +85,7 @@ bibliography: paper.bib
 Fisheries bycatch is one of the top threats to marine mammals worldwide. Data on bycatch and marine mammal abundance are required to make management decisions, yet many marine mammal populations exist in locations where these data are sparse. Population models offer a way to explore the long-term impacts of different bycatch scenarios even when data are sparse or absent. Our modeling tool, `mmrefpoints`, uses very basic information about marine mammal life history, estimates of abundance and bycatch, and population models to project long-term (e.g. 20-100 yr) outcomes of different bycatch rates. These long-term outcomes are the basis of several reference points used for marine mammal management including the Potential Biological Removal (PBR) approach [@wade_calculating_1998]. The goal is to make complex population models accessible to managers and stakeholders who need them, and to students who are learning about risk-based approaches to managing marine mammal populations.
 
 ## Summary
-This tool provides a way for managers and other stakeholders to explore bycatch scenarios, based on simple information about marine mammal life history and rough estimates of abundance and bycatch. The tool consists of an R package and a Shiny application. The primary machinery in the package is an age-structured population dynamics model, which is used to model future population size based on the current population size, life history traits, and bycatch rates. The package also contains tools for calculating performance metrics, as well as the U.S. reference point for bycatch (Potential Biological Removal or PBR) and a solver that estimates the maximum bycatch rate that will meet management objectives. For users who would prefer to see outputs in an interactive user interface, there is a Shiny app called the Marine Mammal Bycatch Impacts Exploration Tool (MMBIET) that shows projections, explains and calculates reference points, and creates a report summarizing inputs and outputs. The app is hosted publicly on the web via the [shinyapps.io server](https://msiple.shinyapps.io/mmrefpoints/), or it can be run locally by installing the package and running the `run_app()` function in the R or RStudio console:
+This tool provides a way for managers and other stakeholders to explore bycatch scenarios, based on simple information about marine mammal life history and rough estimates of abundance and bycatch. The tool consists of an R package [@r2021] and a Shiny application [@shiny2021]. The primary machinery in the package is an age-structured population dynamics model, which is used to model future population size based on the current population size, life history traits, and bycatch rates. The package also contains tools for calculating performance metrics, as well as the U.S. reference point for bycatch (Potential Biological Removal or PBR) and a solver that estimates the maximum bycatch rate that will meet management objectives. For users who would prefer to see outputs in an interactive user interface, there is a Shiny app called the Marine Mammal Bycatch Impacts Exploration Tool (MMBIET) that shows projections, explains and calculates reference points, and creates a report summarizing inputs and outputs. The app is hosted publicly on the web via the [shinyapps.io server](https://msiple.shinyapps.io/mmrefpoints/), or it can be run locally by installing the package and running the `run_app()` function in the R or RStudio console:
 
 ```{r}
 remotes::install_github("mcsiple/mmrefpoints")
@@ -95,7 +95,7 @@ run_app()
 
 ## Population model
 The population model is a simplified version of the model described in Breiwick et al. [-@breiwick_population_1984] and Punt [-@punt_a._e._annex_1999]. The model is a single-sex, age-structured population model. The number of calves or pups born each year is density-dependent, based on the number of mature adults, the pregnancy rate at pre-exploitation equilibrium, maximum theoretical fecundity rate, degree of compensation, and the total abundance relative to carrying capacity K. 
[…]
