+++ paper.bib
@@ -231,3 +231,21 @@
 	pages = {1316},
 	file = {Full Text PDF:/Users/mcsiple/Zotero/storage/CGXWUKSW/Hammond et al. - 2021 - Estimating the Abundance of Marine Mammal Populati.pdf:application/pdf},
 }
+
+@Manual{r2021,
+    title = {R: A Language and Environment for Statistical Computing},
+    author = {{R Core Team}},
+    organization = {R Foundation for Statistical Computing},
+    address = {Vienna, Austria},
+    year = {2021},
+    url = {https://www.R-project.org/},
+}
+
+@Manual{shiny2021,
+    title = {shiny: Web Application Framework for R},
+    author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
+    year = {2021},
+    note = {R package version 1.7.1},
+    url = {https://CRAN.R-project.org/package=shiny},
+}
+
