### 2022-03-07 #3888 @fboehm
@mcsiple - one small fix is needed in the new reference to Shiny: Please fix the bib file so that R is capitalized on line 114 of the pdf. You can accomplish this by placing braces around a capital R, like {R} within the bibtex entry. Thanks again!
