### 2022-03-07 10959e5
Fix R in caps https://github.com/openjournals/joss-reviews/issues/3888
