+++ paper.bib
@@ -242,7 +242,7 @@
 }
 
 @Manual{shiny2021,
-    title = {shiny: Web Application Framework for R},
+    title = {shiny: Web Application Framework for {R}},
     author = {Winston Chang and Joe Cheng and JJ Allaire and Carson Sievert and Barret Schloerke and Yihui Xie and Jeff Allen and Jonathan McPherson and Alan Dipert and Barbara Borges},
     year = {2021},
     note = {R package version 1.7.1},
