### 2022-03-07 #3888 @mcsiple
Fixed! Sorry about that. I copied from `citation` in R and didn't realize I needed that extra formatting to get the caps. Should be good now.
