### 2022-03-08 #3888 @xuanxu
The error is caused because there are two invalid issn values in the bib file: [here](https://github.com/mcsiple/mmrefpoints/blob/master/paper.bib#L69) and [here](https://github.com/mcsiple/mmrefpoints/blob/master/paper.bib#L86).
