### 2022-03-08 5f321b9
Fixed incorrect ISSNs in .bib file https://github.com/openjournals/joss-reviews/issues/3888#issuecomment-1061576880
