+++ paper.bib
@@ -66,7 +66,7 @@
 @article{punt_conserving_2018,
 	title = {Conserving and recovering vulnerable marine species: a comprehensive evaluation of the {US} approach for marine mammals},
 	volume = {75},
-	issn = {1054-3139, 1095-9289},
+	issn = {1054-3139},
 	shorttitle = {Conserving and recovering vulnerable marine species},
 	url = {https://academic.oup.com/icesjms/advance-article/doi/10.1093/icesjms/fsy049/5035891},
 	doi = {10.1093/icesjms/fsy049},
@@ -83,7 +83,7 @@
 @article{breiwick_population_1984,
 	title = {Population {Dynamics} of {Western} {Arctic} {Bowhead} {Whales} ( \textit{{Balaena} mysticetus} )},
 	volume = {41},
-	issn = {0706-652X, 1205-7533},
+	issn = {0706-652X},
 	url = {http://www.nrcresearchpress.com/doi/10.1139/f84-058},
 	doi = {10.1139/f84-058},
 	language = {en},
