+++ CITATION.cff
@@ -45,5 +45,5 @@ authors:
     orcid: https://orcid.org/0000-0002-9776-6605
     
 title: mcsiple/mmrefpoints: Journal submission pre-release
-version: v.1.0.0
-date-released: 2022-3-7
+version: v.1.0.1
+date-released: 2022-3-8
