### 2022-03-10 #3888 @kyleniemeyer
Hi @mcsiple, I've made a small PR for the paper that does some minor formatting changes: https://github.com/mcsiple/mmrefpoints/pull/62

Could you merge this? I will then accept your paper.
