### 2022-03-11 c15af38
Fix Phil's L's in paper.md https://github.com/openjournals/joss-reviews/issues/3888
