+++ README.Rmd
@@ -27,11 +27,11 @@ knitr::opts_chunk$set(
 Margaret C. Siple  
 André E. Punt  
 Tessa B. Francis  
-Phil S. Hammond  
+Philip S. Hammond  
 Dennis Heinemann  
 Kristy J. Long  
 Jeffrey E. Moore  
-Maritza Sepulveda  
+Maritza Sepúlveda  
 Randall R. Reeves  
 Guðjón Már Sigurðsson  
 Gísli Víkingsson  
