### 2022-03-11 #3888 @mcsiple
Upon step 0, I found that one of the authors has an extra L in his name in paper.md🤦🏻‍♀️ It's correct everywhere else and I just fixed it. What do I do next? Sorry for the inconvenience...
