### 2022-03-12 #3888 @mcsiple
Found two more copy edits in paper.md (sorry, and thank you)-- see below.
