### 2022-03-12 07ff128
Two quick copy edits to author names in paper.md https://github.com/openjournals/joss-reviews/issues/3888
