### 2022-03-15 #3888 @arfon
@kyleniemeyer – this should be updated now. Note that the new PDF might take a few hours to show up on the JOSS site due to caching.
